fix: add webui-router to e2e test suite and unify JS package build/test infra
The webui-router package was missing from the xtask e2e test suite, so its
tests never ran in CI. Two architectural contract tests were also silently
broken because they assumed esbuild's compact output format.

What changed:
- Replace esbuild with tsc for library builds in webui-framework and
  webui-router. Libraries should emit per-file ESM for tree-shaking, not
  bundles. esbuild is removed from both packages' devDependencies.
- Replace esbuild with tsc for unit test compilation in both packages.
  Tests only import relative siblings and node built-ins, so bundling was
  unnecessary.
- Remove esbuild from the webui-framework e2e server build. The playwright
  config now runs the server directly via node --experimental-strip-types,
  eliminating the build:e2e step entirely.
- Colocate webui-router unit tests in src/ (e.g. src/inventory.test.ts)
  to match the webui-framework pattern, instead of a separate test/ dir.
  Delete stale test/matcher.test.ts (referenced non-existent src/matcher.ts)
  and test/run-tests.ts (aggregator, replaced by glob pattern).
- Add webui-router as a PlaywrightSuite in xtask/src/e2e.rs so its tests
  run in CI. Relax the suite filter from playwright.config.ts to
  package.json so non-Playwright test suites are included.
- Fix two router tests that inspected handleNavigation source via
  .toString(): widen the abort-gate search window from 100 to 200 chars
  (tsc preserves indentation), and check for 'transition.finished' instead
  of '.finished' (tsc preserves comments that contain that substring).
