+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules/
 dist/
+dist-test/
 bin/
 obj/
 target/
