+++ packages/webui-framework/package.json
@@ -16,22 +16,19 @@
     "dist/"
   ],
   "scripts": {
-    "build": "find src -name '*.ts' ! -name '*.test.ts' | xargs esbuild --outdir=dist --outbase=src --format=esm --platform=browser && tsc --emitDeclarationOnly",
+    "build": "tsc",
     "typecheck": "tsc --noEmit",
     "typecheck:e2e": "tsc -p tsconfig.test.json --noEmit",
-    "build:e2e:scripts": "esbuild ./tests/server.ts --bundle --external:esbuild --outdir=tests/dist --format=esm --platform=node --target=es2022",
-    "build:e2e": "pnpm build:e2e:scripts",
     "test": "pnpm test:unit && pnpm test:e2e",
-    "test:unit": "esbuild \"src/**/*.test.ts\" --bundle --outdir=dist/tests --outbase=src --format=esm --platform=node --target=es2022 && node --test \"dist/tests/**/*.test.js\"",
-    "test:e2e": "pnpm typecheck:e2e && pnpm build:e2e && playwright test",
+    "test:unit": "tsc -p tsconfig.unit-test.json && node --test \"dist/**/*.test.js\"",
+    "test:e2e": "pnpm typecheck:e2e && playwright test",
     "test:update-snapshots": "pnpm test:e2e:update-snapshots",
-    "test:e2e:update-snapshots": "pnpm typecheck:e2e && pnpm build:e2e && playwright test --update-snapshots"
+    "test:e2e:update-snapshots": "pnpm typecheck:e2e && playwright test --update-snapshots"
   },
   "devDependencies": {
     "@microsoft/webui-test-support": "workspace:*",
     "@playwright/test": "catalog:",
     "@types/node": "catalog:",
-    "esbuild": "catalog:",
     "typescript": "catalog:"
   }
 }
