+++ packages/webui-framework/tsconfig.json
@@ -11,8 +11,6 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
     "declaration": true,
-    "declarationDir": "dist",
-    "emitDeclarationOnly": true,
     "outDir": "dist",
     "rootDir": "src"
   },
