--- packages/webui-router/test/run-tests.ts
@@ -1,5 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import './navigation-path.test.ts';
-import './matcher.test.ts';
