+++ packages/webui-router/tsconfig.unit-test.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "declaration": false,
+    "types": ["urlpattern-polyfill", "dom-navigation", "dom-view-transitions", "node"]
+  },
+  "exclude": ["node_modules", "dist"]
+}
